+++ index.js
@@ -1,3 +1,4 @@
 export { default as NonLocalStorage } from './dist/esm/NonLocalStorage.js'
+export { default as createOfflineNonLocalStorage } from './dist/esm/OfflineNonLocalStorage.js'
 export { default as createSyncObject } from './dist/esm/SyncObject.js'
 export { default as createOfflineSyncObject } from './dist/esm/OfflineSyncObject.js'