+++ index.ts
@@ -1,4 +1,5 @@
 export { default as NonLocalStorage } from './dist/deno/NonLocalStorage.ts'
+export { default as createOfflineNonLocalStorage } from './dist/deno/OfflineNonLocalStorage.ts'
 export { default as createSyncObject } from './dist/deno/SyncObject.ts'
 export { default as createOfflineSyncObject } from './dist/deno/OfflineSyncObject.ts'
 export type * from './dist/deno/types.ts'